Fix for HTTP Error 403 Forbidden
This fix was applied for mainly downloading playlists from youtube
music.
Each time the error is raised it will store the position in playlist at
which the download was interrupted by the error. And it will launch
again in a recursion.
The way it does this, is through recursion in the wrapper function
within function `__handle_extraction_exceptions`